third_party/WebKit/LayoutTests/http/tests/xmlhttprequest/set-dangerous-headers-expected.txt

CONSOLE WARNING: line 10: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
CONSOLE ERROR: line 12: Refused to set unsafe header "ACCEPT-CHARSET"
CONSOLE ERROR: line 13: Refused to set unsafe header "ACCEPT-ENCODING"
CONSOLE ERROR: line 14: Refused to set unsafe header "ACCESS-CONTROL-REQUEST-HEADERS"
CONSOLE ERROR: line 15: Refused to set unsafe header "ACCESS-CONTROL-REQUEST-METHOD"
CONSOLE ERROR: line 21: Refused to set unsafe header "CONNECTION"
CONSOLE ERROR: line 22: Refused to set unsafe header "CONTENT-LENGTH"
CONSOLE ERROR: line 23: Refused to set unsafe header "COOKIE"
CONSOLE ERROR: line 24: Refused to set unsafe header "COOKIE2"
CONSOLE ERROR: line 25: Refused to set unsafe header "DATE"
CONSOLE ERROR: line 26: Refused to set unsafe header "DNT"
CONSOLE ERROR: line 27: Refused to set unsafe header "EXPECT"
CONSOLE ERROR: line 28: Refused to set unsafe header "HOST"
CONSOLE ERROR: line 29: Refused to set unsafe header "KEEP-ALIVE"
CONSOLE ERROR: line 30: Refused to set unsafe header "ORIGIN"
CONSOLE ERROR: line 31: Refused to set unsafe header "REFERER"
CONSOLE ERROR: line 32: Refused to set unsafe header "TE"
CONSOLE ERROR: line 33: Refused to set unsafe header "TRAILER"
CONSOLE ERROR: line 34: Refused to set unsafe header "TRANSFER-ENCODING"
CONSOLE ERROR: line 35: Refused to set unsafe header "UPGRADE"
CONSOLE ERROR: line 36: Refused to set unsafe header "USER-AGENT"
CONSOLE ERROR: line 37: Refused to set unsafe header "VIA"
CONSOLE ERROR: line 39: Refused to set unsafe header "Proxy-"
CONSOLE ERROR: line 40: Refused to set unsafe header "Proxy-test"
CONSOLE ERROR: line 41: Refused to set unsafe header "PROXY-FOO"
CONSOLE ERROR: line 43: Refused to set unsafe header "Sec-"
CONSOLE ERROR: line 44: Refused to set unsafe header "Sec-test"
CONSOLE ERROR: line 45: Refused to set unsafe header "SEC-FOO"
Test that setRequestHeader cannot be used to alter security-sensitive headers.

SUCCESS
